Modularised pipeline into different files to make it a bit easier to read, will work on main to get pipeline to work
