--- UI.html
@@ -1,11 +1,0 @@
-<html>
-<head>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" href="static/styles.css">
-    <title>Chatbot UI</title>
-</head>
-
-<body>
-    <h1 class = "header"> Hitachi Chatbot UI</h1>
-</body>
-</html>
