+++ pdf_folder/ENSC3016_Course_Notes_Part_1_Electromagnetism_Transformers.pdf
